# app/static/openapi.yaml
openapi: 3.0.1
servers:
- url: https://resources.operationcode.org/api/v1
  description: Default Server
info:
  title: Operation Code Resources API
  version: '1.0'
  x-logo:
    url: 'https://operation-code-assets.s3.us-east-2.amazonaws.com/branding/logos/small-blue-logo.png'
    altText: 'Operation Code Logo'
    backgroundColor: '#252e3e'
    href: 'https://operationcode.org/contact'
  description: |
    # Introduction
    Operation Code Resources API is a collection of programming and cyber security learning resources curated by and for the Operation Code community.

    # Authentication
    Create (POST) and update (PUT) routes are protected and require an API key. API key requests to the `/apikey` endpoint require valid Operation Code credentials in the request body.

    If you are not already a member of our community, join Operation Code [HERE](https://operationcode.org/join).

    When you create a request to an authenticated route, you *MUST* include the header `x-apikey` along with your API key.

    # Versioning
    To ensure contracts with the api are not broken with future updates, you may supply a version number in your request header.
    Include an `X-Api-Version` header with a valid version in your request:
      ```
      X-Api-Version: 1.0
      ```
    ### Default Versions
...


  contact:
    name: Operation Code
    url: https://operationcode.org/contact
paths:
...
  /resources/{id}:
    get:
      summary: Get Resource
      description: Gets more details about a specific resource, given the resource ID
      tags:
        - Get Resource
      parameters:
      - name: id
        in: path
        description: Resource ID
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: Resource details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
              example:
                apiVersion: '1.0'
                resource:
                  category: 'Books'
                  created_at: '2019-06-15 17:55:32'
                  downvotes: 0
                  id: 3
                  languages: []
                  last_updated: '2019-06-15 17:55:32'
                  name: 'Free Tech Books'
                  notes: 'Focuses on general computer science concepts rather than a specific language'
                  free: true
                  times_clicked: 0
                  upvotes: 0
                  url: 'http://www.freetechbooks.com/'
                  user_vote_direction: null
                status: 'ok'
                status_code: 200
        404:
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update Resource
      description: Updates an existing learning resource entry. The request body should be a JSON object containing the properties to be updated and their corresponding new values.
      tags:
        - Update Resource
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceRequest'
            example:
              name: 'Updated Book Title'
        required: true
      parameters:
        - name: id
          in: path
          description: Resource ID
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Updated resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
              example:
                apiVersion: '1.0'
                resource:
                  category: 'Tutorial'
                  created_at: '2019-07-21 15:41:19'
                  downvotes: 0
                  id: 2137
                  languages: []
                  last_updated: '2019-07-22 16:05:53'
                  name: 'Updated Book Title'
                  notes: null
                  free: true
                  times_clicked: 3
                  upvotes: 1
                  url: 'http://www.test.com'
                  user_vote_direction: null
                status: 'ok'
                status_code: 200
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/UnauthorizedNoApiKey'
        422:
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                apiVersion: '1.0'
                errors:
                  invalid-params:
                      message: 'The following params were invalid: url. Resource id 2137 already has this URL.'
                      params:
                        - 'url'
                      resource: 'https://resources.operationcode.org/api/v1/2137'
                status: 'Unprocessable Entity'
                status_code: 422
      security:
      - ApiKeyAuth: []
...
  /resources:
    get:
      summary: List Resources
      description: Gets list of all available resource entries or filtered entries based on provided criteria. Note&#58; The `page` attribute in the response of the `/search` endpoint begins with `0`, but the `page` attribute in the response of the `List Resources` endpoint begins with `1`.
      tags:
        - List Resources
      parameters:
        - in: query
          name: free
          required: false
          description: Whether the resource is free or not. To search for *free* resources, make a `GET` request to `/resources?free=true`.
          schema:
            type: boolean
        - in: query
          name: languages
          required: false
          description: Language(s) of the resource. For example, to filter for Python and JavaScript resources, make a `GET` request to `/resources?languages=python&languages=javascript`
          schema:
            type: string
        - in: query
          name: category
          required: false
          description: Category of resource. For example, to filter for book resources, make a `GET` request to `/resources?category=books`.
          schema:
            type: string
        - in: query
          name: updated_after
          required: false
          description: Filters resources that were updated after the specified date. To get all resources updated after July 17, 2019, use `/resources?updated_after=2019-07-17`
          schema:
            type: string
            format: date-time

      responses:
        200:
          description: A list of resources
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilteredResourcesResponse'
              example:
                apiVersion: '1.0'
                resources:
                  - category': 'Tutorial'
                    created_at: '2019-07-21 15:41:19'
                    downvotes: 0
                    id: 2137
                    languages: []
                    last_updated: '2019-07-22 17:39:50'
                    name: 'Some, but not ALL of the knowledge'
                    notes: null
                    free: true
                    times_clicked: 3
                    upvotes: 1
                    url: 'http://www.test.com'
                    user_vote_direction: null
                has_next: false
                has_prev: false
                number_of_pages: 1
                page: 1
                records_per_page: 20
                status: 'ok'
                status_code: 200
                total_count: 1
        404:
          $ref: '#/components/responses/NotFound'

        422:
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                apiVersion: '1.0'
                errors:
                  unprocessable-entity:
                    message: 'The value for "updated_after" is invalid'
                status: 'Unprocessable Entity'
                status_code: 422

    post:
      summary: Create Resources
      description: Creates new learning resources. The structure of the request is a list of resource objects. The required properties for each new resource are `category`, `name`, `free`, and `url`.
      tags:
        - Create Resources
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResourcesRequest'
            example:
              - category: 'Tutorial'
                name: 'CSS Tutorial'
                free: true
                url: 'https://www.w3schools.com/css/'
        required: true
      responses:
        200:
          description: Created resources
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceList'
              example:
                apiVersion: '1.0'
                resources:
                  - category: 'Tutorial'
                    created_at: '2019-07-21 15:41:19'
                    downvotes: 0
                    id: 2137
                    languages:
                      - 'CSS'
                    last_updated: '2019-07-22 16:05:53'
                    name: 'CSS Tutorial'
                    notes: null
                    free: true
                    times_clicked: 3
                    upvotes: 1
                    url: 'https://www.w3schools.com/css/'
                status: 'ok'
                status_code: 200
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/UnauthorizedNoApiKey'
        422:
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                  apiVersion: '1.0'
                  errors:
                    - index: 0
                      invalid-params:
                        message: 'The following params were invalid: url. Resource id 2137 already has this URL.'
                        params:
                          - url
                        resource: 'https://resources.operationcode.org/api/v1/2137'
                      missing-params:
                          message: 'The following params were missing: name, category, free.'
                          params:
                            - name
                            - category
                            - free
                  status: 'Unprocessable Entity'
                  status_code: 422
      security:
      - ApiKeyAuth: []
components:
  schemas:
    Error:
      type: object
      properties:
        apiVersion:
          type: string
          description: API version
        errors:
          $ref: '#/components/schemas/ErrorData'
        status:
          type: string
          description: Status message
        status_code:
          type: integer
          description: Status code
    ErrorData:
      type: object
      description: The response errors
      properties:
        missing-params:
          type: object
          properties:
            params:
              description: The associated parameters.
              type: array
              items:
                type: string
            message:
              description: Error message
              type: string
        invalid-params:
          type: object
          properties:
            params:
              description: The associated parameters.
              type: array
              items:
                type: string
            message:
              description: Error message
              type: string
            resource:
              description: An href link to the resource with a duplicate url
              type: string
        unauthorized:
          type: object
          properties:
            message:
              description: Error message
              type: string
        missing-body:
          type: object
          properties:
            message:
              description: Error message
              type: string
        unprocessable-entity:
          type: object
          properties:
            message:
              description: Error message
              type: string
        bad-request:
          type: object
          properties:
            message:
              description: Error message
              type: string
        not-found:
          type: object
          properties:
            message:
              description: Error message
              type: string
    ErrorResponse:
      type: object
      properties:
        apiVersion:
          type: string
          description: API version
        errors:
          $ref: '#/components/schemas/ErrorDataList'
        status:
          type: string
          description: Status message
        status_code:
          type: integer
          description: Status code
    ErrorDataList:
      type: array
      items:
        type: object
        properties:
          index:
            type: integer
            description: The index of the invalid resource
          missing-params:
            type: object
            properties:
              params:
                description: The associated parameters.
                type: array
                items:
                  type: string
              message:
                description: Error message
                type: string
          invalid-params:
            type: object
            properties:
              params:
                description: The associated parameters.
                type: array
                items:
                  type: string
              message:
                description: Error message
                type: string
              resource:
                description: An href link to the resource with a duplicate url
                type: string
          unauthorized:
            type: object
            properties:
              message:
                description: Error message
                type: string
          missing-body:
            type: object
            properties:
              message:
                description: Error message
                type: string
          unprocessable-entity:
            type: object
            properties:
              message:
                description: Error message
                type: string
          bad-request:
            type: object
            properties:
              message:
                description: Error message
                type: string
          not-found:
            type: object
            properties:
              message:
                description: Error message
                type: string
    Resource:
      type: object
      properties:
        apiVersion:
          type: string
          description: API version
        resource:
          $ref: '#/components/schemas/ResourceData'
        status:
          type: string
          description: Status message
        status_code:
          type: integer
          description: Status code
...
    CreateResourcesRequest:
      allOf:
        - $ref: '#/components/schemas/ResourcesRequest'
        # - required: [category, name, free, url]
    ResourcesRequest:
      type: array
      items:
        type: object
        properties:
          category:
            type: string
            description: Category
          languages:
            type: array
            description: List of languages
            items:
              type: string
          name:
            type: string
            description: Resource name
          notes:
            type: string
            description: Resource notes
          free:
            type: boolean
            description: True or false for whether resource is free
          url:
            type: string
            description: Resource url
    ResourceRequest:
      type: object
      properties:
        category:
          type: string
          description: Category
        languages:
          type: array
          description: List of languages
          items:
            type: string
        name:
          type: string
          description: Resource name
        notes:
          type: string
          description: Resource notes
        free:
          type: boolean
          description: True or false for whether resource is free
        url:
          type: string
          description: Resource url
    ResourceData:
      type: object
      properties:
        category:
          type: string
          description: Category
        created_at:
          type: string
          format: date-time
          description: Date created
        downvotes:
          type: integer
          description: Negative ratings
        id:
          type: integer
          description: Resource ID
        last_updated:
          type: string
          format: date-time
          description: Date updated
        languages:
          type: array
          description: List of languages
          items:
            type: string
        name:
          type: string
          description: Resource name
        notes:
          type: string
          description: Resource notes
        free:
          type: boolean
          description: True or false for whether resource is free
        upvotes:
          type: integer
          description: Positive ratings
        url:
          type: string
          description: Resource url
    ResourceList:
      type: array
      items:
        $ref: '#/components/schemas/ResourceData'
...
    FilteredResourcesResponse:
      type: object
      properties:
        apiVersion:
          type: string
          description: API version
        resources:
          type: array
          items:
            $ref: '#/components/schemas/ResourceData'
        has_next:
          type: boolean
          description: Additional pages
        has_previous:
          type: boolean
          description: Previous pages
        number_of_pages:
          type: integer
          description: Total pages
        page:
          type: integer
          description: Current page
        records_per_page:
          type: integer
          description: Records per page
        status:
          type: string
          description: Status message
        status_code:
          type: integer
          description: Status code
        total_count:
          type: integer
          description: Total resources retrieved

##############################
# API Request and Response
##############################
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            apiVersion: '1.0'
            errors:
              not-found:
                message: 'The resource you requested does not exist.'
            status: 'Not Found'
            status_code: 404
...
    UnauthorizedNoApiKey:
      description: Unauthorized access
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            apiVersion: '1.0'
            errors:
              unauthorized:
                message: 'You must provide a valid API token in the x-apikey header.'
            status: 'Unauthorized'
            status_code: 401
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            apiVersion: '1.0'
            errors:
              bad-request:
                message: "400 Bad Request: Failed to decode JSON object: Expecting ',' delimiter: line 3 column 2 (char 25)"
            status: 'Bad Request'
            status_code: 400
  securitySchemes:
    ApiKeyAuth:
      description: API key authentication
      type: apiKey
      name: x-apikey
      in: header

##################
# Groups and Tags
##################
...